Add function, change format of query return value
The function added is `submit_if_unarchived`, which first checks if url
hasn't been archived yet. Then, it proceeds to archive the url.
Otherwise, it returns the query result.

The format of the return value of `query_wayback` is now a list of
dicts, where the keys are the corresponding field names returned by the
API.

All relevant functions now raise an error when the HTTP response is an
error.