Change return value of function, cdz->cdx
Instead of `submit_if_unarchived` returning a Response object, it
returns a tuple of url and the contents of the archived site.
The endpoint used previously ends with cdz, while the
official docs say cdx. I tested it and it turns out that it doesn't
matter (it could basically be any string), so it uses cdx instead.