Abstract extract_urls to strings instead of fnames
`extract_urls` now takes a string as input. `extract_urls_from_file` is
a new function which takes over the previous job of `extract_urls`, and
it simply uses `extract_urls` to do so. These changes will enable the
script to use pipe inputs as the input, instead of only text files.